More MainMonitor: Fixed --help & segfault on quit First: Fixed the --help parsing so that we properly check for the program name. argv[0] is the full path for the command if not run from $PATH, so we now check against the object->GetName().
Second:
We were segfaulting on quit when running on a real socket due to a
malformed if check against deleting the poll2_server object. (NOTE: that
due to how the client/server is implemented here the actual poll2
program is the client in the scenario)

Third: I hagve removed redundant header includes from the
mainmonitor.hpp. These are included via the monitor.hpp and therefore do
not need to be double included here.