Features and Fixes for the monitors
Colors and Rows are now working for both main and sub monitors. 

Color threshold structs are defined for nSHE, vandle, and mtas, in addition to the default one. Currently vandle, mtas and the default have the same values but this will be changed when i confirm where folks want the colors.